Bootstraps clean-up and project refactor
