--- client/src/components/About.js
@@ -1,10 +1,0 @@
-const About = () => {
-    return (
-        <>
-        <h1>About Page</h1>
-        <p>Version 0</p>
-        </>
-    );
-}
-
-export default About;
