+++ client/src/views/About/About.js
@@ -1,0 +1,16 @@
+import { Container, Row, Col } from "react-bootstrap";
+
+const About = () => {
+    return (
+        <Container>
+            <Row>
+                <Col md={12}> 
+                    <h1>About</h1>
+                    <p>Version 0</p>
+                </Col>
+            </Row>
+        </Container>
+    );
+}
+
+export default About;
